+++ Sources/Sentry/SentryAutoBreadcrumbTrackingIntegration.m
@@ -3,6 +3,9 @@
 #import "SentryDefaultCurrentDateProvider.h"
 #import "SentryDependencyContainer.h"
 #import "SentryFileManager.h"
+#import "SentryLog.h"
+#import "SentryOptions+Private.h"
+#import "SentryOptions.h"
 #import "SentrySystemEventBreadcrumbs.h"
 
 NS_ASSUME_NONNULL_BEGIN
@@ -19,6 +22,13 @@ SentryAutoBreadcrumbTrackingIntegration ()
 
 - (void)installWithOptions:(nonnull SentryOptions *)options
 {
+    if (!options.enableAutoBreadcrumbTracking) {
+        [SentryLog logWithMessage:@"Not going to enable BreadcrumbTracking because "
+                                  @"enableAutoBreadcrumbTracking is disabled."
+                         andLevel:kSentryLevelDebug];
+        [options removeEnabledIntegration:NSStringFromClass([self class])];
+        return;
+    }
 
     [self installWithOptions:options
              breadcrumbTracker:[[SentryBreadcrumbTracker alloc]