+++ Sources/Sentry/include/SentryAutoBreadcrumbTrackingIntegration.h
@@ -1,5 +1,4 @@
 #import "SentryIntegrationProtocol.h"
-#import <Foundation/Foundation.h>
 
 NS_ASSUME_NONNULL_BEGIN